updtaed route.ts + authOptions.ts
