+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,88 +1,5 @@
-/* eslint-disable no-param-reassign */
-
-import NextAuth, { DefaultSession, NextAuthOptions, Session } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import { PrismaClient } from '@prisma/client';
-import bcrypt from 'bcrypt';
-import { JWT } from 'next-auth/jwt';
-
-const prisma = new PrismaClient();
-
-declare module 'next-auth' {
-  interface Session extends DefaultSession {
-    user: {
-      id: string;
-      role: string;
-    } & DefaultSession['user'];
-  }
-
-  interface User {
-    id: string;
-    role: string;
-  }
-}
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: { label: 'Email', type: 'text' },
-        password: { label: 'Password', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          return null;
-        }
-
-        try {
-          const user = await prisma.user.findUnique({
-            where: { email: credentials.email },
-          });
-
-          if (!user) return null;
-
-          const passwordMatch = await bcrypt.compare(credentials.password, user.password);
-          if (!passwordMatch) return null;
-
-          return {
-            id: user.id.toString(),
-            email: user.email,
-            role: user.role,
-          };
-        } catch (error) {
-          console.error('Authentication error:', error);
-          return null;
-        }
-      },
-    }),
-  ],
-  pages: {
-    signIn: '/auth/signin',
-    signOut: '/auth/signout',
-    error: '/auth/error',
-  },
-  session: {
-    strategy: 'jwt',
-    maxAge: 30 * 24 * 60 * 60, // 30 days
-  },
-  callbacks: {
-    async jwt({ token, user }: { token: JWT; user?: { id: string; role: string } }) {
-      if (user) {
-        token.id = user.id;
-        token.role = user.role;
-      }
-      return token;
-    },
-    async session({ session, token }: { session: Session; token: JWT }) {
-      if (session.user) {
-        session.user.id = token.id as string;
-        session.user.role = token.role as string;
-      }
-      return session;
-    },
-  },
-};
+import NextAuth from 'next-auth';
+import { authOptions } from '@/lib/authOptions';
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
