+++ src/app/company-profile/page.tsx
@@ -5,7 +5,7 @@ import { prisma } from '@/lib/prisma';
 import { getServerSession } from 'next-auth';
 import { Job } from '@prisma/client';
 import { loggedInProtectedPage } from '@/lib/page-protection';
-import authOptions from '@/lib/authOptions';
+import { authOptions } from '@/lib/authOptions';
 import JobListings from '@/components/JobListings';
 import AddJobForm from '@/components/AddJobForm';
