+++ src/app/student-home-page/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/authOptions';
 import prisma from '../../../prisma/prisma';
 
 const StudentHomePage = async () => {
